Feat/a4 (#7)
* Finish reachability analysis.

* Finish A4

* Ensure that two CSTNodes are == only if they're reference equal.

* Fix bugs and tests.

Inside getAccessibleInstanceMembers and getAccessibleStaticMembers, if Object was not in IOD, we'd enter an infinite loop. We were using == to compare primite types, which doesn't work because == is a reference equality check on CSTNode objects. In LocalVariableDeclarations, you could use a variable in its own initializer after it was initialized. This isn't possible for fields. So, I added a test for fields, and implemented the behaviour for variables.

* Make Environment.pushScope and Environment.popScope private and instead rely on Environment.withNewScope inside NameResolver

* Make CSTNode a regular class, so we don't have to use id to hack == into a reference equality check.